+++ Nb.Tolls.Infrastructure/Models/TollFeesModel.cs
@@ -4,7 +4,7 @@ public class TollFeesModel
 {
     public required string Start { get; init; }
     public required string End { get; init; }
-    public required decimal AmountSek { get; init; }
+    public required int AmountSek { get; init; }
     public required int StartMin { get; init; }
     public required int EndMin { get; init; }
 }
